+++ src/main/java/ch/admin/bit/jeap/opensearch/registry/verifier/ValidationResult.java
@@ -13,16 +13,22 @@ import java.util.stream.Stream;
 @Data
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 public class ValidationResult {
-    private static final ValidationResult OK = new ValidationResult(true, Collections.emptyList());
+    private static final ValidationResult OK = new ValidationResult(
+            true, Collections.emptyList(), Collections.emptyList());
     private final boolean valid;
     private final List<String> errors;
+    private final List<String> warnings;
 
     public static ValidationResult ok() {
         return OK;
     }
 
     public static ValidationResult fail(String error) {
-        return new ValidationResult(false, Collections.singletonList(error));
+        return new ValidationResult(false, Collections.singletonList(error), Collections.emptyList());
+    }
+
+    public static ValidationResult warning(String warning) {
+        return new ValidationResult(true, Collections.emptyList(), Collections.singletonList(warning));
     }
 
     public static ValidationResult merge(ValidationResult... results) {
@@ -32,6 +38,8 @@ public class ValidationResult {
     private static ValidationResult merge(ValidationResult r1, ValidationResult r2) {
         List<String> errors = Stream.concat(r1.getErrors().stream(), r2.getErrors().stream())
                 .collect(Collectors.toList());
-        return new ValidationResult(r1.isValid() && r2.isValid(), errors);
+        List<String> warnings = Stream.concat(r1.getWarnings().stream(), r2.getWarnings().stream())
+                .collect(Collectors.toList());
+        return new ValidationResult(r1.isValid() && r2.isValid(), errors, warnings);
     }
 }